requirements.txt
pandas
numpy
tqdm
python-chess
psycopg2-binary==2.8.5
python-dotenv
tensorflow-gpu
keras
seaborn